server iterates through towers differently each time
If a bunch towers are fighting, then the tower whose splatters get
added to the array last has an advantage because it will overwrite the
other splatters. To counteract this, the server doesn’t iterate through
towers in the same order every time, rather it iterates through them
round robin style and for each “round” in the round robin, it will
iterate through them forwards and then backwards. I dunno if this is
perfectly fair, but it’s definitely better.